requirements.txt
atlassian-python-api>=3.41.14
requests>=2.32.3
python-dotenv>=1.0.1